fix: Resolve remaining compilation errors in collectors, DTOs, and repositories
FinanceMetricsDto:
- Add unsettledRestaurantAmount, unsettledCourierAmount to PayoutSummaryDto
- Add discountUsageRate to DiscountSummaryDto
- Add approvedRefunds, pendingRefunds, rejectedRefunds to RefundSummaryDto

DashboardFilterRequest:
- Add pageSize field and setPageSize method for builder compatibility

CxSummaryDto:
- Add restaurantAverageRating, courierAverageRating, appStoreAverageRating aliases

FraudMetricsRequest:
- Add alias getters for threshold values (velocityThreshold, refundRateThreshold, etc.)

BehavioralFraudMetricsDto:
- Add anomalousOrderCount to OrderValueAnomaliesDto with getter
- Add highRefundUsers, refundAbusersList aliases to RefundAbuseMetricsDto

ReferralFraudMetricsDto:
- Add overallReferralFraudRate, totalFraudulentReferrals aliases
- Add circularReferralMetrics, promoAbuseMetrics aliases with getters

SecurityLogMetricsDto:
- Add failedLoginRate alias to LoginSecurityMetricsDto
- Add totalRateLimitViolations to RateLimitMetricsDto
- Add suspiciousIpCount, suspiciousIpsList to IpSecurityMetricsDto
- Add detectedAttacks to BruteForceMetricsDto

FraudSummaryDto:
- Add totalEstimatedLoss to FinancialImpactDto

BackendResourceMetricsDto:
- Change connectedClients from Integer to Long

SmsRestTemplateConfig:
- Fix setConnectTimeout/setReadTimeout method names

DashboardRestaurantRepository:
- Add countTotalRestaurants, countBusyRestaurants, countOfflineRestaurants
- Add countTemporarilyClosedRestaurants
- Add avgPreparationTime, avgOrderAcceptanceLatency, avgRestaurantRating
- Add avgOrderAcceptanceRate
- Add findRestaurantDetailsFiltered, findAllRestaurantDetails
- Add countByCuisineType, countByCity
- Add findTopPerformingRestaurants, findUnderperformingRestaurants

RestaurantMetricsDto:
- Add busyRestaurants field
- Add avgPreparationTimeMinutes, rejectedOrdersCount, rejectionRate to summary

RestaurantCollector:
- Fix primitive int null comparison using getPageNumber()/getPageSize()

+++ src/main/java/com/fooddelivery/analytics/technical/dto/BackendResourceMetricsDto.java
@@ -313,7 +313,7 @@ public class BackendResourceMetricsDto {
         /**
          * Connected clients.
          */
-        private Integer connectedClients;
+        private Long connectedClients;
 
         /**
          * Used memory in bytes.
